Ajout conditions pour lancer script et fonction usage; ajout log exemple fail2ban pour test

+++ banlist_temp.sh
@@ -1,7 +1,30 @@
 #!/bin/bash
 #
+# ban2block script
+#
 ip='[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}'
 file='fail2ban.log*'
+log='banned.txt'
+#
+usage() {
+  cat<<EOF
+
+This script will check fail2ban.log and block all IPs, already banned more than X times, with iptables
+All IP blocked will be logged into $log for further usage.
+
+REQUIREMENTS:
+  - root priviledge
+  - file /var/log/fail2ban.log
+
+Usage: no arguments needed on this version.
+
+EOF
+}
+# A few test before to start
+# Need to be roor to use this script
+if [ `id -u` -ne 0 ]; then usage; exit 1; fi
+if [ ! -f $file ]; then usage; exit 1; fi
+if [ ! -f $log ]; then touch $log; fi
 #
 grep -o -E "Ban $ip" $file | grep -o -E "$ip" > banned.tmp
 cat banned.tmp | sort | uniq > sort-banned.tmp
@@ -34,5 +57,6 @@ while read line
     # Process iptables blocking
     # iptables -I INPUT 1 -s $k -j DROP
     echo $k >> iptables_ip.txt
+    echo $k >> banned.txt
     echo "[OK] - $k has just been blocked."
   done < fin-results.txt
